requirements.txt
numpy==2.4.4
cupy-cuda12x>=13.0.0